+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 
 ## [1.0.6] - 2024-0x-xx
 
+### Changed
+
+- 2024-01-09 Adding log level, can be changed from settings
+
 ## [1.0.5] - 2024-01-08
 
 ### Fixed